requirement.txt
gradio==4.44.0
python-dotenv==1.0.1
google-generativeai==0.7.1
json2